--- JaeWook/CodeTest/CodeTest/CodeTest.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.34114.132
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeTest", "CodeTest\CodeTest.csproj", "{CDE59608-BE4F-4430-8939-0ECC04DAAAEB}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MainCode", "..\..\MainCode\MainCode.csproj", "{1EA22B88-3492-4544-89F3-1EA7370CF4A8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CDE59608-BE4F-4430-8939-0ECC04DAAAEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CDE59608-BE4F-4430-8939-0ECC04DAAAEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CDE59608-BE4F-4430-8939-0ECC04DAAAEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CDE59608-BE4F-4430-8939-0ECC04DAAAEB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1EA22B88-3492-4544-89F3-1EA7370CF4A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1EA22B88-3492-4544-89F3-1EA7370CF4A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1EA22B88-3492-4544-89F3-1EA7370CF4A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1EA22B88-3492-4544-89F3-1EA7370CF4A8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CA01A869-F4F6-42E0-90E6-A8447FD9A680}
-	EndGlobalSection
-EndGlobal
