--- JaeWook/CodeTest/CodeTest/CodeTest/HelloWorld.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MainCode
-{
-    class HelloWorld
-    {
-        static void Main(string[] args)
-        {
-            Console.WriteLine("Hello World!");
-
-
-        }
-
-    }
-}
