--- JaeWook/CodingTest/CodingTest.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.34114.132
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodingTest", "CodingTest\CodingTest.csproj", "{852B8440-D2EA-4944-A297-43C16CED48E2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{852B8440-D2EA-4944-A297-43C16CED48E2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{852B8440-D2EA-4944-A297-43C16CED48E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{852B8440-D2EA-4944-A297-43C16CED48E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{852B8440-D2EA-4944-A297-43C16CED48E2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {79E32598-C0A6-4759-9102-7B630CD15F9F}
-	EndGlobalSection
-EndGlobal
