--- JaeWook/CodingTest/CodingTest/HelloWorld.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-//위대한 시작, 안녕 세상아
-namespace CodingTest
-{
-    class HelloWorld
-    {
-        
-    }
-}
