--- JaeWook/commenttest
@@ -1,4 +1,0 @@
-test for comment 1
-test for comment 2
-test for comment 3
-test for comment 4
